ST-662: Removed unwanted code and fixed few sonar issues
