+++ src/main/java/uk/gov/hmcts/sptribs/CaseApiApplication.java
@@ -16,7 +16,7 @@ import uk.gov.hmcts.reform.ccd.client.CoreCaseDataClientAutoConfiguration;
 import uk.gov.hmcts.reform.idam.client.IdamApi;
 import uk.gov.hmcts.sptribs.document.DocAssemblyClient;
 import uk.gov.hmcts.sptribs.document.DocumentManagementClient;
-import uk.gov.hmcts.sptribs.hearingvenue.LocationClient;
+import uk.gov.hmcts.sptribs.recordlisting.LocationClient;
 import uk.gov.hmcts.sptribs.payment.FeesAndPaymentsClient;
 import uk.gov.hmcts.sptribs.payment.PaymentPbaClient;
 import uk.gov.hmcts.sptribs.systemupdate.service.ScheduledTaskRunner;
