--- src/main/java/uk/gov/hmcts/sptribs/hearingvenue/model/HearingVenueResponse.java
@@ -1,22 +1,0 @@
-package uk.gov.hmcts.sptribs.hearingvenue.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class HearingVenueResponse {
-
-    @JsonProperty("hearingVenues")
-    private List<HearingVenue> hearingVenues;
-
-}
