+++ src/main/java/uk/gov/hmcts/sptribs/recordlisting/LocationClient.java
@@ -1,25 +1,21 @@
-package uk.gov.hmcts.sptribs.hearingvenue;
+package uk.gov.hmcts.sptribs.recordlisting;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.GetMapping;
-import uk.gov.hmcts.sptribs.hearingvenue.model.HearingVenue;
-import uk.gov.hmcts.sptribs.hearingvenue.model.HearingVenueResponse;
-import uk.gov.hmcts.sptribs.hearingvenue.model.Region;
-import uk.gov.hmcts.sptribs.hearingvenue.model.RegionResponse;
-
-import java.util.Optional;
+import uk.gov.hmcts.sptribs.recordlisting.model.HearingVenue;
+import uk.gov.hmcts.sptribs.recordlisting.model.Region;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static uk.gov.hmcts.sptribs.common.config.ControllerConstants.SERVICE_AUTHORIZATION;
-import static uk.gov.hmcts.sptribs.hearingvenue.HearingVenueConstants.IS_CASE_MANAGEMENT_LOCATION;
-import static uk.gov.hmcts.sptribs.hearingvenue.HearingVenueConstants.IS_HEARING_LOCATION;
-import static uk.gov.hmcts.sptribs.hearingvenue.HearingVenueConstants.IS_TEMPORARY_LOCATION;
-import static uk.gov.hmcts.sptribs.hearingvenue.HearingVenueConstants.LOCATION_TYPE;
-import static uk.gov.hmcts.sptribs.hearingvenue.HearingVenueConstants.REGION_ID;
-import static uk.gov.hmcts.sptribs.hearingvenue.HearingVenueConstants.REGION_ID_1;
+import static uk.gov.hmcts.sptribs.recordlisting.RecordListingConstants.IS_CASE_MANAGEMENT_LOCATION;
+import static uk.gov.hmcts.sptribs.recordlisting.RecordListingConstants.IS_HEARING_LOCATION;
+import static uk.gov.hmcts.sptribs.recordlisting.RecordListingConstants.IS_TEMPORARY_LOCATION;
+import static uk.gov.hmcts.sptribs.recordlisting.RecordListingConstants.LOCATION_TYPE;
+import static uk.gov.hmcts.sptribs.recordlisting.RecordListingConstants.REGION_ID;
+import static uk.gov.hmcts.sptribs.recordlisting.RecordListingConstants.REGION_ID_1;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
