+++ src/main/java/uk/gov/hmcts/sptribs/recordlisting/RecordListingConstants.java
@@ -1,6 +1,6 @@
-package uk.gov.hmcts.sptribs.hearingvenue;
+package uk.gov.hmcts.sptribs.recordlisting;
 
-public class HearingVenueConstants {
+public class RecordListingConstants {
 
     public static final String REGION_ID = "region_id";
     public static final String REGION_ID_1 = "regionId";
@@ -10,6 +10,6 @@ public class HearingVenueConstants {
     public static final String IS_TEMPORARY_LOCATION = "is_temporary_location";
     public static final String HYPHEN = "-";
 
-    private HearingVenueConstants() {
+    private RecordListingConstants() {
     }
 }
