+++ src/main/java/uk/gov/hmcts/sptribs/recordlisting/model/HearingVenue.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.hearingvenue.model;
+package uk.gov.hmcts.sptribs.recordlisting.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,7 +6,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 
 import java.time.LocalDate;
